+++ ICD.Connect.Routing.Mock/Switcher/MockSwitcherDeviceSettings.cs
@@ -1,22 +1,10 @@
-﻿using System;
-using ICD.Connect.Devices;
+﻿using ICD.Connect.Devices;
 using ICD.Connect.Settings.Attributes;
 
 namespace ICD.Connect.Routing.Mock.Switcher
 {
-	[KrangSettings(FACTORY_NAME)]
+	[KrangSettings("MockSwitcherDevice", typeof(MockSwitcherDevice))]
 	public sealed class MockSwitcherDeviceSettings : AbstractDeviceSettings
 	{
-		private const string FACTORY_NAME = "MockSwitcherDevice";
-
-		/// <summary>
-		/// Gets the originator factory name.
-		/// </summary>
-		public override string FactoryName { get { return FACTORY_NAME; } }
-
-		/// <summary>
-		/// Gets the type of the originator for this settings instance.
-		/// </summary>
-		public override Type OriginatorType { get { return typeof(MockSwitcherDevice); } }
 	}
 }