+++ ICD.Connect.Routing/Endpoints/Destinations/AbstractDestinationSettings.cs
@@ -2,11 +2,5 @@
 {
 	public abstract class AbstractDestinationSettings : AbstractSourceDestinationBaseSettings, IDestinationSettings
 	{
-		private const string DESTINATION_ELEMENT = "Destination";
-
-		/// <summary>
-		/// Gets the xml element.
-		/// </summary>
-		protected override string Element { get { return DESTINATION_ELEMENT; } }
 	}
 }