+++ ICD.Connect.Routing/Endpoints/Sources/AbstractSourceSettings.cs
@@ -2,11 +2,5 @@
 {
 	public abstract class AbstractSourceSettings : AbstractSourceDestinationBaseSettings, ISourceSettings
 	{
-		private const string SOURCE_ELEMENT = "Source";
-
-		/// <summary>
-		/// Gets the xml element.
-		/// </summary>
-		protected override string Element { get { return SOURCE_ELEMENT; } }
 	}
 }