+++ ICD.Connect.Routing/Proxies/AbstractProxyRouteDestinationControl.cs
@@ -8,7 +8,14 @@ namespace ICD.Connect.Routing.Proxies
 {
 	public abstract class AbstractProxyRouteDestinationControl : AbstractProxyRouteControl, IProxyRouteDestinationControl
 	{
+		/// <summary>
+		/// Raised when an input source status changes.
+		/// </summary>
 		public event EventHandler<SourceDetectionStateChangeEventArgs> OnSourceDetectionStateChange;
+
+		/// <summary>
+		/// Raised when the device starts/stops actively using an input, e.g. unroutes an input.
+		/// </summary>
 		public event EventHandler<ActiveInputStateChangeEventArgs> OnActiveInputsChanged;
 
 		/// <summary>