+++ ICD.Connect.Routing/RoutingGraphs/AbstractRoutingGraph.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using ICD.Common.Utils;
 using ICD.Common.Utils.Services;
 using ICD.Connect.API.Commands;
+using ICD.Connect.API.Nodes;
 using ICD.Connect.Routing.Connections;
 using ICD.Connect.Routing.ConnectionUsage;
 using ICD.Connect.Routing.Controls;
@@ -13,7 +12,6 @@ using ICD.Connect.Routing.Endpoints.Groups;
 using ICD.Connect.Routing.Endpoints.Sources;
 using ICD.Connect.Routing.EventArguments;
 using ICD.Connect.Routing.StaticRoutes;
-using ICD.Connect.Routing.Utils;
 using ICD.Connect.Settings;
 
 namespace ICD.Connect.Routing.RoutingGraphs
@@ -21,19 +19,61 @@ namespace ICD.Connect.Routing.RoutingGraphs
 	public abstract class AbstractRoutingGraph<TSettings> : AbstractOriginator<TSettings>, IRoutingGraph
 		where TSettings : IRoutingGraphSettings, new()
 	{
+		/// <summary>
+		/// Raised when a route operation fails or succeeds.
+		/// </summary>
 		public abstract event EventHandler<RouteFinishedEventArgs> OnRouteFinished;
+
+		/// <summary>
+		/// Raised when a switcher changes routing.
+		/// </summary>
 		public abstract event EventHandler<SwitcherRouteChangeEventArgs> OnRouteChanged;
+
+		/// <summary>
+		/// Raised when a source device starts/stops sending video.
+		/// </summary>
 		public abstract event EventHandler<EndpointStateEventArgs> OnSourceTransmissionStateChanged;
+
+		/// <summary>
+		/// Raised when a source device is connected or disconnected.
+		/// </summary>
 		public abstract event EventHandler<EndpointStateEventArgs> OnSourceDetectionStateChanged;
+
+		/// <summary>
+		/// Raised when a destination device changes active input state.
+		/// </summary>
 		public abstract event EventHandler<EndpointStateEventArgs> OnDestinationInputActiveStateChanged;
 
 		#region Properties
 
+		/// <summary>
+		/// Gets the connections collection.
+		/// </summary>
 		public abstract IConnectionsCollection Connections { get; }
+
+		/// <summary>
+		/// Gets the connection usages collection.
+		/// </summary>
 		public abstract IConnectionUsageCollection ConnectionUsages { get; }
+
+		/// <summary>
+		/// Gets the static routes collection.
+		/// </summary>
 		public abstract IOriginatorCollection<StaticRoute> StaticRoutes { get; }
+
+		/// <summary>
+		/// Gets the sources collection.
+		/// </summary>
 		public abstract ISourceCollection Sources { get; }
+
+		/// <summary>
+		/// Gets the destinations collection.
+		/// </summary>
 		public abstract IDestinationCollection Destinations { get; }
+
+		/// <summary>
+		/// Gets the destination groups collection.
+		/// </summary>
 		public abstract IOriginatorCollection<IDestinationGroup> DestinationGroups { get; }
 
 		/// <summary>
@@ -98,13 +138,13 @@ namespace ICD.Connect.Routing.RoutingGraphs
 		/// </summary>
 		/// <param name="destination"></param>
 		/// <param name="input"></param>
-		/// <param name="type"></param>
+		/// <param name="flag"></param>
 		/// <param name="signalDetected">When true skips inputs where no video is detected.</param>
 		/// <param name="inputActive"></param>
 		/// <exception cref="ArgumentNullException"></exception>
 		/// <returns>The source</returns>
 		public abstract EndpointInfo? GetActiveSourceEndpoint(IRouteDestinationControl destination, int input,
-		                                                      eConnectionType type,
+		                                                      eConnectionType flag,
 		                                                      bool signalDetected, bool inputActive);
 
 		/// <summary>
@@ -647,6 +687,39 @@ namespace ICD.Connect.Routing.RoutingGraphs
 		#region Console
 
 		/// <summary>
+		/// Gets the child console nodes.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<IConsoleNodeBase> GetConsoleNodes()
+		{
+			foreach (IConsoleNodeBase node in GetBaseConsoleNodes())
+				yield return node;
+
+			foreach (IConsoleNodeBase node in RoutingGraphConsole.GetConsoleNodes(this))
+				yield return node;
+		}
+
+		/// <summary>
+		/// Wrokaround for "unverifiable code" warning.
+		/// </summary>
+		/// <returns></returns>
+		private IEnumerable<IConsoleNodeBase> GetBaseConsoleNodes()
+		{
+			return base.GetConsoleNodes();
+		}
+
+		/// <summary>
+		/// Calls the delegate for each console status item.
+		/// </summary>
+		/// <param name="addRow"></param>
+		public override void BuildConsoleStatus(AddStatusRowDelegate addRow)
+		{
+			base.BuildConsoleStatus(addRow);
+
+			RoutingGraphConsole.BuildConsoleStatus(this, addRow);
+		}
+
+		/// <summary>
 		/// Gets the child console commands.
 		/// </summary>
 		/// <returns></returns>
@@ -655,22 +728,8 @@ namespace ICD.Connect.Routing.RoutingGraphs
 			foreach (IConsoleCommand command in GetBaseConsoleCommands())
 				yield return command;
 
-			yield return
-				new ConsoleCommand("PrintTable", "Prints a table of the routed devices and their input/output information.",
-				                   () => PrintTable());
-			yield return new ConsoleCommand("PrintConnections", "Prints the list of all connections.", () => PrintConnections());
-			yield return new ConsoleCommand("PrintSources", "Prints the list of Sources", () => PrintSources());
-			yield return new ConsoleCommand("PrintDestinations", "Prints the list of Destinations", () => PrintDestinations());
-			yield return new ConsoleCommand("PrintUsages", "Prints a table of the connection usages.", () => PrintUsages());
-
-			yield return new GenericConsoleCommand<int, int, eConnectionType, int>("Route",
-			                                                                       "Routes source to destination. Usage: Route <sourceId> <destId> <connType> <roomId>",
-			                                                                       (a, b, c, d) =>
-				                                                                       RouteConsoleCommand(a, b, c, d));
-			yield return new GenericConsoleCommand<int, int, eConnectionType, int>("RouteGroup",
-			                                                                       "Routes source to destination group. Usage: Route <sourceId> <destGrpId> <connType> <roomId>",
-			                                                                       (a, b, c, d) =>
-				                                                                       RouteGroupConsoleCommand(a, b, c, d));
+			foreach (IConsoleCommand command in RoutingGraphConsole.GetConsoleCommands(this))
+				yield return command;
 		}
 
 		/// <summary>
@@ -680,110 +739,6 @@ namespace ICD.Connect.Routing.RoutingGraphs
 		private IEnumerable<IConsoleCommand> GetBaseConsoleCommands()
 		{
 			return base.GetConsoleCommands();
-		}
-
-		private string PrintSources()
-		{
-			TableBuilder builder = new TableBuilder("Id", "Source");
-
-			foreach (ISource source in Sources.GetChildren().OrderBy(c => c.Id))
-				builder.AddRow(source.Id, source);
-
-			return builder.ToString();
-		}
-
-		private string PrintDestinations()
-		{
-			TableBuilder builder = new TableBuilder("Id", "Destination");
-
-			foreach (IDestination destination in Destinations.GetChildren().OrderBy(c => c.Id))
-				builder.AddRow(destination.Id, destination);
-
-			return builder.ToString();
-		}
-
-		/// <summary>
-		/// Loop over the devices, build a table of inputs, outputs, and their statuses.
-		/// </summary>
-		private string PrintTable()
-		{
-			RoutingGraphTableBuilder builder = new RoutingGraphTableBuilder(this);
-			return builder.ToString();
-		}
-
-		private string PrintConnections()
-		{
-			TableBuilder builder = new TableBuilder("Source", "Output", "Destination", "Input", "Type");
-
-			foreach (Connection con in Connections.GetChildren().OrderBy(c => c.Source.Device).ThenBy(c => c.Source.Address))
-				builder.AddRow(con.Source, con.Source.Address, con.Destination, con.Destination.Address, con.ConnectionType);
-
-			return builder.ToString();
-		}
-
-		/// <summary>
-		/// Loop over the connections and build a table of usages.
-		/// </summary>
-		private string PrintUsages()
-		{
-			TableBuilder builder = new TableBuilder("Connection", "Type", "Source", "Rooms");
-
-			Connection[] connections = Connections.ToArray();
-
-			for (int index = 0; index < connections.Length; index++)
-			{
-				Connection connection = connections[index];
-				ConnectionUsageInfo info = ConnectionUsages.GetConnectionUsageInfo(connection);
-				int row = 0;
-
-				foreach (eConnectionType type in EnumUtils.GetFlagsExceptNone(connection.ConnectionType))
-				{
-					string connectionString = row == 0 ? string.Format("{0} - {1}", connection.Id, connection.Name) : string.Empty;
-					EndpointInfo? source = info.GetSource(type);
-					int[] rooms = info.GetRooms(type).ToArray();
-					string roomsString = rooms.Length == 0 ? string.Empty : StringUtils.ArrayFormat(rooms);
-
-					builder.AddRow(connectionString, type, source, roomsString);
-
-					row++;
-				}
-
-				if (index < connections.Length - 1)
-					builder.AddSeparator();
-			}
-
-			return builder.ToString();
-		}
-
-		private string RouteConsoleCommand(int source, int destination, eConnectionType connectionType, int roomId)
-		{
-			if (!Sources.ContainsChild(source) || !Destinations.ContainsChild(destination))
-				return "Krang does not contains a source or destination with that id";
-
-			Route(Sources.GetChild(source), Destinations.GetChild(destination), connectionType, roomId);
-
-			return "Sucessfully executed route command";
-		}
-
-		private string RouteGroupConsoleCommand(int source, int destination, eConnectionType connectionType, int roomId)
-		{
-			if (!Sources.ContainsChild(source) || !DestinationGroups.ContainsChild(destination))
-				return "Krang does not contains a source or destination group with that id";
-
-			Route(Sources.GetChild(source), DestinationGroups.GetChild(destination), connectionType, roomId);
-
-			return "Sucessfully executed route command";
-		}
-
-		private void Route(ISource source, IDestinationGroup destinationGroup, eConnectionType connectionType, int roomId)
-		{
-			foreach (
-				IDestination destination in
-				destinationGroup.Destinations.Where(Destinations.ContainsChild).Select(d => Destinations.GetChild(d)))
-			{
-				IDestination destination1 = destination;
-				ThreadingUtils.SafeInvoke(() => Route(source, destination1, connectionType, roomId));
-			}
 		}
 
 		#endregion