+++ ICD.Connect.Routing/RoutingGraphs/IRoutingGraph.cs
@@ -47,16 +47,34 @@ namespace ICD.Connect.Routing.RoutingGraphs
 
 		#region Properties
 
+		/// <summary>
+		/// Gets the connections collection.
+		/// </summary>
 		IConnectionsCollection Connections { get; }
 
+		/// <summary>
+		/// Gets the connection usages collection.
+		/// </summary>
 		IConnectionUsageCollection ConnectionUsages { get; }
 
+		/// <summary>
+		/// Gets the static routes collection.
+		/// </summary>
 		IOriginatorCollection<StaticRoute> StaticRoutes { get; }
 
+		/// <summary>
+		/// Gets the sources collection.
+		/// </summary>
 		ISourceCollection Sources { get; }
 
+		/// <summary>
+		/// Gets the destinations collection.
+		/// </summary>
 		IDestinationCollection Destinations { get; }
 
+		/// <summary>
+		/// Gets the destination groups collection.
+		/// </summary>
 		IOriginatorCollection<IDestinationGroup> DestinationGroups { get; }
 
 		#endregion
@@ -93,13 +111,13 @@ namespace ICD.Connect.Routing.RoutingGraphs
 		/// </summary>
 		/// <param name="destination"></param>
 		/// <param name="input"></param>
-		/// <param name="type"></param>
+		/// <param name="flag"></param>
 		/// <param name="signalDetected">When true skips inputs where no video is detected.</param>
 		/// <param name="inputActive"></param>
 		/// <exception cref="ArgumentNullException"></exception>
 		/// <returns>The source</returns>
 		EndpointInfo? GetActiveSourceEndpoint(IRouteDestinationControl destination, int input,
-		                                      eConnectionType type, bool signalDetected, bool inputActive);
+		                                      eConnectionType flag, bool signalDetected, bool inputActive);
 
 		/// <summary>
 		/// Finds the actively routed source for the destination at the given input address.